corrección de tipo de dato a la tabla producto

+++ programa/CRM/CRM/Models/Producto.cs
@@ -13,7 +13,7 @@ namespace CRM.Models
         public int Codigo { get; set; }
         public int CodigoFamilia { get; set; }
         public string? Nombre { get; set; }
-        public int? PrecioEstandar { get; set; }
+        public double? PrecioEstandar { get; set; }
         public int? Estado { get; set; }
         public string? Descripcion { get; set; }
